+++ ctutil/ct_ssl.c
@@ -39,7 +39,7 @@ ct_sha1(uint8_t *src, uint8_t *dst, size_t len)
 {
 	SHA_CTX		ctx;
 
-	CDBG("src %p len %zu", src, len);
+	CDBG("src %p len %lu", src, (unsigned long) len);
 
 	SHA1_Init(&ctx);
 	SHA1_Update(&ctx, src, len);
@@ -55,7 +55,7 @@ ct_sha1_setup(SHA_CTX *ctx)
 void
 ct_sha1_add(uint8_t *src, SHA_CTX *ctx, size_t len)
 {
-	CDBG("src %p len %zu", src, len);
+	CDBG("src %p len %lu", src, (unsigned long) len);
 
 	SHA1_Update(ctx, src, len);
 }
@@ -84,7 +84,7 @@ ct_sha512(uint8_t *src, uint8_t *dst, size_t len)
 {
 	SHA512_CTX		ctx;
 
-	CDBG("src %p len %zu", src, len);
+	CDBG("src %p len %lu", src, (unsigned long) len);
 
 	SHA512_Init(&ctx);
 	SHA512_Update(&ctx, src, len);
